move ResourceNotFoundException to internal.exception package
