+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/io/StaticResolver.java
@@ -25,8 +25,8 @@ import org.wso2.carbon.uuf.core.App;
 import org.wso2.carbon.uuf.core.Component;
 import org.wso2.carbon.uuf.core.Theme;
 import org.wso2.carbon.uuf.exception.FileOperationException;
-import org.wso2.carbon.uuf.exception.ResourceNotFoundException;
 import org.wso2.carbon.uuf.internal.UUFServer;
+import org.wso2.carbon.uuf.internal.exception.ResourceNotFoundException;
 import org.wso2.carbon.uuf.internal.io.util.MimeMapper;
 import org.wso2.carbon.uuf.spi.HttpRequest;
 import org.wso2.carbon.uuf.spi.HttpResponse;
